Add files via upload
Fixed a slough of bugs. Now collumns and rows correlate correctly in the file. Moved the board creation to a class board, through which everything can be managed. Level generation should be implemented as member functions of the Board class now and conditionally called as opposed to directly in the code. You can now proceed to the next level by locating the red box found in the center of one of the rooms on the map.

Thoughs for expansion:
-Make the maze scale with level (Something like collumns,rows = board.level + randint(1,4)  or something of the like to allow progression
-Add cover objects which are randomly placed in rooms to give diversity to the rooms contents
-Create enemies to slay in the maze
-Create a gating mechanism to prevent progression till a room is cleared
-Add GUI stuff to show the level you're currently on
-Make the player able to shoot - bullets should collide with walls/enemies and be destroyed and be destroyed after travel_time
     -You can use a try bloc