Add files via upload
Some behind the scenes cleaning up - remove functions now exist for environment, wall, and goal objects and they now remove from drawing instead of hiding them away.
Implemented levels visually - rooms now get assigned a level (for use to spawn baddies, etc) and the level of floor you are on is now displayed in the top right. The room level is currently displayed in the middle of the room to show it works and is accessable in the main loop.
Note on polygons: Pygame only accounts for rectangle collision, circular collision and polygon collision all will have to be implemented unless you want to import another library (called Box2D). Circles should be relatively simple, however polygons are going to have to use point collision.
Another option is to draw the units as polygons and have hitboxes that are rectangles.